install/requirements.txt
flask==3.1.2
python-dotenv==1.2.1
requests==2.32.4
urllib3==2.6.2
werkzeug==3.1.3
pillow==12.0.0
pi-heif==1.1.1
pytz==2025.2
openai==2.14.0
google-genai>=1.64.0
numpy>=2.2.2
icalendar==6.3.1
recurring-ical-events==3.7.0
psutil==7.0.0
cysystemd==2.0.1
waitress==3.0.2
feedparser==6.0.11
astral>=3.1
sounddevice
scipy
shazamio
ai-edge-litert
skyfield>=1.45
sgp4
